Support polymophism when sending or receiving interfaces
