+++ src/Umbraco.Infrastructure/Serialization/IUmbracoJsonTypeInfoResolver.cs
@@ -1,0 +1,8 @@
+using System.Text.Json.Serialization.Metadata;
+
+namespace Umbraco.Cms.Infrastructure.Serialization;
+
+public interface IUmbracoJsonTypeInfoResolver : IJsonTypeInfoResolver
+{
+    IEnumerable<Type> FindSubTypes(Type type);
+}
